+++ src/test_typing_extensions/utils.py
@@ -1,0 +1,5 @@
+# We just import the utils module to check for syntax or import problems.
+import typing_extensions.utils  # noqa F401
+
+def dummy():
+    pass